+++ src/languages/services/yaccHover.ts
@@ -1,22 +1,31 @@
 import { TextDocument, Hover, Position, MarkedString, MarkdownString } from 'vscode';
-import { YACCDocument, ISymbol } from "../parser/yaccParser";
+import { YACCDocument, ISymbol, NodeType } from "../parser/yaccParser";
 import { createMarkedCodeString } from "./utils";
 
 export function doYACCHover(document: TextDocument, position: Position, yaccDocument: YACCDocument): Hover | null {
     const offset = document.offsetAt(position);
-    const node = yaccDocument.getEmbeddedNode(offset);
-    if (node) {
+    const code = yaccDocument.getEmbeddedNode(offset);
+    if (code) {
         return null;
     }
 
-    const word = document.getText(document.getWordRangeAtPosition(position));
-
-    var message: MarkdownString | undefined = undefined;
     var symbol: ISymbol;
-    if ((symbol = yaccDocument.types[word])) {
-        message = createMarkedCodeString(`${symbol.type} ${symbol.name}`, 'yacc');
+    const word = document.getText(document.getWordRangeAtPosition(position));
+    const node = yaccDocument.getNodeByOffset(offset);
+    if (node) {
+        // Inside <...>
+        if (node.typeOffset && offset > node.typeOffset) {
+            if (!node.typeEnd || offset <= node.typeEnd) {
+                if ((symbol = yaccDocument.types[word])) {
+                    message = createMarkedCodeString(symbol.type, 'yacc');
+                    return { contents: [createMarkedCodeString(symbol.type, 'yacc')] }
+                }
+                return null;
+            }
+        }
     }
 
+    var message: MarkdownString | undefined = undefined;
     if ((symbol = yaccDocument.symbols[word])) {
         message = createMarkedCodeString(`%type <${symbol.type ? symbol.type : '?'}> ${symbol.name}`, 'yacc');
     }