+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## UnReleased
 
+## 0.2.1
+### Added
+- Added support for initial state scope (lex)
+
+### Changed
+- Now uses a parser to parse the %union types instead of regex (yacc)
+
+### Bug fixes
+- Fixed wrong string highlight when quotes are escaped (lex)
+- Fixed wrong type detection in the %union scope (yacc)
+- Fixed an hover issue when the non-terminal name is the same as a typename (yacc) 
+
 ## Release
 
 ## 0.2.0 27/04/2020